# openapi.yaml
openapi: 3.1.0
info:
  title: Family Hub API
  version: "1.0.0"
  description: |
    Family Hub backend API (Express + TypeScript + Supabase/Postgres).

    This spec documents the API **as it actually behaves today**, including
    two genuinely different response-envelope conventions and three
    different auth conventions, depending on which route file an endpoint
    lives in. See `docs/API.md` for a human-oriented companion/index to this
    spec, including narrative explanation of the quirks below.

    ## Response envelopes (two incompatible conventions)

    - **Newer routes** (`chores`, `learning`, `announcements`, `reminders`,
      `energy`, `calendar`, `family`, `activity-log`, `smartthings`) return
      an object with a top-level `status: 'success' | 'error'` field and a
      `timestamp` ISO string on every response. On error, the shape is
      `{ status: 'error', message: string, error?: string }`
      (see `EnvelopeError`). On success, most — but **not all** — of these
      routes nest the payload under a `data` key alongside `count?`
      (see `EnvelopeSuccess`). Several endpoints in this same "newer" group
      (`chores`, `smartthings`, `learning`, `performance`-adjacent routes)
      instead use an endpoint-specific top-level key (e.g. `chores`,
      `devices`, `stats`, `progress`, `activity`, `history`) instead of
      `data`. Each such endpoint documents its literal response shape
...
      path.
...

    ## Auth conventions (three, genuinely different)
...
servers:
...
security: []
paths:
...
  /api/users:
    get:
      tags: [Users]
      summary: List all users (admin)
      operationId: getAllUsers
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Users list.
          content:
            application/json:
              schema:
                type: object
                properties:
                  count: { type: integer }
                  users:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string }
                        name: { type: string }
                        email: { type: string }
                        role: { type: string }
                        created_at: { type: string, format: date-time }
        '401':
          description: Missing authorization header.
          content: { application/json: { schema: { $ref: '#/components/schemas/LegacyErrorBody' } } }
        '500':
          description: Unexpected failure.
          content: { application/json: { schema: { $ref: '#/components/schemas/LegacyErrorBody' } } }

  # ================================================================
  # BADGES
  # ================================================================
  /api/badges:
    get:
      tags: [Badges]
      summary: List all available badges
      operationId: listBadges
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: All badges.
          content:
            application/json:
              schema:
                type: object
                properties:
                  count: { type: integer }
                  badges:
                    type: array
                    items: { $ref: '#/components/schemas/Badge' }
        '401':
          description: Missing authorization header.
          content: { application/json: { schema: { $ref: '#/components/schemas/LegacyErrorBody' } } }
        '500':
          description: Unexpected failure.
          content: { application/json: { schema: { $ref: '#/components/schemas/LegacyErrorBody' } } }
...
  /api/chores:
    get:
      tags: [Chores]
      summary: List chores with today's completion state
      description: >-
        `scope=mine` (default) — the caller's enabled chores, for the board.
        `scope=family` — every family member's chores including disabled ones,
        with the assignee name, for the parent Manage panel.
      operationId: listChores
      security: [{ userIdHeader: [] }]
      parameters:
        - name: scope
          in: query
          schema: { type: string, enum: [mine, family], default: mine }
      responses:
        '200':
          description: Chores.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, const: success }
                  chores:
                    type: array
                    items: { $ref: '#/components/schemas/ChoreWithStatus' }
                  count: { type: integer }
                  timestamp: { type: string, format: date-time }
        '401':
          description: Missing x-user-id.
          content: { application/json: { schema: { $ref: '#/components/schemas/EnvelopeError' } } }
        '500':
          description: Unexpected failure.
          content: { application/json: { schema: { $ref: '#/components/schemas/EnvelopeError' } } }
    post:
      tags: [Chores]
      summary: Create a chore
      description: >-
        `assigneeId` defaults to the caller and is stored as the chore's
        `userId` ("whose chore"); it must be an active member of the caller's
        family.
      operationId: createChore
      security: [{ userIdHeader: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, timeSlot, pointsValue]
              properties:
                name: { type: string }
                description: { type: string }
                timeSlot: { type: string, enum: [morning, afternoon, evening] }
                pointsValue: { type: number, minimum: 1 }
                assigneeId: { type: string, format: uuid }
      responses:
        '201':
          description: Chore created.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, const: success }
                  message: { type: string, const: "Chore created successfully" }
                  chore: { $ref: '#/components/schemas/Chore' }
                  timestamp: { type: string, format: date-time }
        '400':
          description: Missing/invalid fields, or `assigneeId` is not a family member.
          content: { application/json: { schema: { $ref: '#/components/schemas/EnvelopeError' } } }
        '401':
          description: Missing x-user-id.
          content: { application/json: { schema: { $ref: '#/components/schemas/EnvelopeError' } } }
        '500':
          description: Unexpected failure.
          content: { application/json: { schema: { $ref: '#/components/schemas/EnvelopeError' } } }
...
  /api/habits:
    get:
      tags: [Habits]
      summary: List habits with this week's progress and streak
      description: >-
        `scope=mine` (default) — the caller's enabled habits, for the board.
        `scope=family` — every family member's habits including disabled ones,
        with the assignee name, for the parent Manage panel.
      operationId: listHabits
      security: [{ userIdHeader: [] }]
      parameters:
        - name: scope
          in: query
          schema: { type: string, enum: [mine, family], default: mine }
      responses:
        '200':
          description: Habits.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, const: success }
                  habits:
                    type: array
                    items: { $ref: '#/components/schemas/HabitWithStatus' }
                  count: { type: integer }
                  timestamp: { type: string, format: date-time }
        '401':
          description: Missing x-user-id.
          content: { application/json: { schema: { $ref: '#/components/schemas/EnvelopeError' } } }
        '500':
          description: Unexpected failure.
          content: { application/json: { schema: { $ref: '#/components/schemas/EnvelopeError' } } }
    post:
      tags: [Habits]
      summary: Create a habit
      description: >-
        `assigneeId` defaults to the caller and is stored as the habit's
        `userId`; it must be an active member of the caller's family.
      operationId: createHabit
      security: [{ userIdHeader: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, weeklyTarget, pointsValue]
              properties:
                title: { type: string }
                description: { type: string }
                weeklyTarget: { type: integer, minimum: 1, maximum: 7 }
                pointsValue: { type: number, minimum: 1 }
                assigneeId: { type: string, format: uuid }
      responses:
        '201':
          description: Habit created.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, const: success }
                  message: { type: string, const: "Habit created successfully" }
                  habit: { $ref: '#/components/schemas/Habit' }
                  timestamp: { type: string, format: date-time }
        '400':
          description: Missing/invalid fields, or `assigneeId` is not a family member.
          content: { application/json: { schema: { $ref: '#/components/schemas/EnvelopeError' } } }
        '401':
          description: Missing x-user-id.
          content: { application/json: { schema: { $ref: '#/components/schemas/EnvelopeError' } } }
        '500':
          description: Unexpected failure.
          content: { application/json: { schema: { $ref: '#/components/schemas/EnvelopeError' } } }
...
  /api/announcements:
    get:
      tags: [Announcements]
      summary: Get announcements visible to the authenticated user's family
      operationId: listAnnouncements
      security: [{ userIdHeader: [] }]
      responses:
        '200':
          description: Announcements.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/EnvelopeSuccess'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/AnnouncementWithReadStatus' }
        '401':
          description: Missing x-user-id.
          content: { application/json: { schema: { $ref: '#/components/schemas/EnvelopeError' } } }
        '500':
          description: Unexpected failure.
          content: { application/json: { schema: { $ref: '#/components/schemas/EnvelopeError' } } }
    post:
      tags: [Announcements]
      summary: Create an announcement
      operationId: createAnnouncement
      security: [{ userIdHeader: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [family_id, title, message]
              properties:
                family_id: { type: string, format: uuid }
                title: { type: string }
                message: { type: string }
                announcement_type: { type: string, default: general }
                priority: { type: string, default: normal }
                target_audience: { type: string, default: all }
                target_user_ids: { type: array, items: { type: string, format: uuid } }
                is_pinned: { type: boolean, default: false }
                expires_at: { type: string, format: date-time }
      responses:
        '201':
          description: Created.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/EnvelopeSuccess'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Announcement' }
        '400':
          description: Missing required fields.
          content: { application/json: { schema: { $ref: '#/components/schemas/EnvelopeError' } } }
        '401':
          description: Missing x-user-id.
          content: { application/json: { schema: { $ref: '#/components/schemas/EnvelopeError' } } }
        '500':
          description: Unexpected failure.
          content: { application/json: { schema: { $ref: '#/components/schemas/EnvelopeError' } } }
...
  /api/meals:
    get:
      tags: [Meals]
      summary: The caller's family meal plan across a date range (default rolling 7 days from today)
      operationId: listMealPlan
      security: [{ userIdHeader: [] }]
      parameters:
        - name: start
          in: query
          required: false
          schema: { type: string, format: date }
        - name: end
          in: query
          required: false
          schema: { type: string, format: date }
      responses:
        '200':
          description: Meal plan slot rows.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/EnvelopeSuccess'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/MealPlanEntry' }
                      range:
                        type: object
                        properties:
                          start: { type: string, format: date }
                          end: { type: string, format: date }
        '400':
          description: start/end not YYYY-MM-DD.
          content: { application/json: { schema: { $ref: '#/components/schemas/EnvelopeError' } } }
        '401':
          description: Missing x-user-id.
          content: { application/json: { schema: { $ref: '#/components/schemas/EnvelopeError' } } }
        '500':
          description: Unexpected failure.
          content: { application/json: { schema: { $ref: '#/components/schemas/EnvelopeError' } } }
...
  /api/reminders:
    get:
      tags: [Reminders]
      summary: List every reminder in the caller's family
      description: >-
        Family-scoped. Each row carries `assignee_name`. Stale recurring reminders
        are rolled forward to their next occurrence as a side effect of this read
        (there is no job runner yet).
      operationId: listReminders
      security: [{ userIdHeader: [] }]
      parameters:
        - name: filter
          in: query
          schema: { type: string, enum: [pending, dismissed, all], default: all }
      responses:
        '200':
          description: Reminders (empty array if the caller has no family).
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/EnvelopeSuccess'
                  - type: object
                    properties:
                      count: { type: integer }
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/Reminder' }
        '401':
          description: Missing x-user-id.
          content: { application/json: { schema: { $ref: '#/components/schemas/EnvelopeError' } } }
        '500':
          description: Unexpected failure.
          content: { application/json: { schema: { $ref: '#/components/schemas/EnvelopeError' } } }
    post:
      tags: [Reminders]
      summary: Create a reminder
      description: >-
        The reminder is created in the caller's family. `assignee_user_id`
        defaults to the caller; anyone in the family may be assigned.
      operationId: createReminder
      security: [{ userIdHeader: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, scheduled_time]
              properties:
                title: { type: string }
                description: { type: string }
                reminder_type: { type: string, enum: [chore, assignment, event, goal, custom], default: custom }
                assignee_user_id:
                  type: string
                  format: uuid
                  description: Family member the reminder is for. Defaults to the caller.
                scheduled_time: { type: string, format: date-time }
                recurrence: { type: string, enum: [once, daily, weekly, monthly], default: once }
                recurrence_end_date: { type: string, format: date }
                related_item_id:
                  type: string
                  description: Links the reminder to another item (e.g. a calendar event id, or its series recurringEventId).
                related_item_type:
                  type: string
                  description: What related_item_id points at, e.g. "calendar_event".
                remind_before_minutes:
                  type: integer
                  minimum: 0
                  description: Lead time before the source item; the reminder still fires at scheduled_time, this is for display / de-dupe.
      responses:
        '201':
          description: Created.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/EnvelopeSuccess'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Reminder' }
        '400':
          description: Invalid body, or `assignee_user_id` is not a member of the caller's family.
          content: { application/json: { schema: { $ref: '#/components/schemas/EnvelopeError' } } }
        '401':
          description: Missing x-user-id.
          content: { application/json: { schema: { $ref: '#/components/schemas/EnvelopeError' } } }
        '404':
          description: The caller has no family.
          content: { application/json: { schema: { $ref: '#/components/schemas/EnvelopeError' } } }
        '500':
          description: Unexpected failure.
          content: { application/json: { schema: { $ref: '#/components/schemas/EnvelopeError' } } }
...
  /api/calendar/auth/google/callback:
    get:
      tags: [Calendar]
      summary: Google OAuth callback (redirect, not JSON)
      description: |
        On success or failure, this always issues a 302 redirect back to
        `${FRONTEND_URL}/dashboard` with a `googleAuth=success` or
        `googleAuth=error&message=...` query param — it never returns a JSON
        body. Also reachable via the legacy alias `GET /auth/google/callback`
        (registered directly in `server.ts`), which itself just 302-redirects
        to this path with the query string preserved.
      operationId: googleOAuthCallback
      parameters:
        - name: code
          in: query
          required: true
          schema: { type: string }
        - name: state
          in: query
          required: true
          schema: { type: string, description: "The user ID, passed through as OAuth state." }
      responses:
        '302':
          description: Redirect to frontend dashboard with googleAuth=success|error.
          headers:
            Location:
              schema: { type: string }
        '400':
          description: Missing code or state (JSON, not a redirect, in this one case).
          content: { application/json: { schema: { $ref: '#/components/schemas/EnvelopeError' } } }
  /auth/google/callback:
    get:
      tags: [Calendar]
      summary: Legacy OAuth callback alias (redirect shim only)
      description: "Registered directly in server.ts. Immediately 302-redirects to /api/calendar/auth/google/callback with the query string forwarded unchanged. Exists for backwards compatibility with a previously-registered Google OAuth redirect URI."
      operationId: legacyGoogleOAuthCallbackAlias
      parameters:
        - name: code
          in: query
          schema: { type: string }
        - name: state
          in: query
          schema: { type: string }
      responses:
        '302':
          description: Redirect to /api/calendar/auth/google/callback.
          headers:
            Location:
              schema: { type: string }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        `Authorization: Bearer <token>` must be present. Used by `auth`,
        `users`, `badges`, `points`, `external-apis`. For most routes in
        this group the server only checks the header is present and
        prefixed `Bearer ` — it does not cryptographically verify the
        token (the one exception is `GET /auth/me`, which does call
        `supabase.auth.getUser(token)`).
    userIdHeader:
      type: apiKey
      in: header
      name: x-user-id
      description: |
        Custom header identifying the caller by user ID. Used by `chores`,
        `learning`, `announcements`, `reminders`, `energy`, `calendar`,
        `family`, `activity-log`. There is no cryptographic verification —
        any UUID-shaped (or even arbitrary) value is trusted as-is. A few
        endpoints within these route files (e.g. `GET /api/energy/usage`,
        `POST /api/activity/log`, several PATCH/DELETE-by-id endpoints in
        `announcements`/`reminders`/`calendar`) do not actually check this
        header despite living in the same route file — see the per-endpoint
        `security` field and description for exceptions.
    kioskTokenHeader:
      type: apiKey
      in: header
      name: x-kiosk-token
      description: |
        Raw device token issued by `POST /api/kiosk/enroll`, presented by a
        registered wall display on `POST /api/kiosk/session` to boot into its
        household with no user login. The server sha256-hashes it and matches
        against `kiosk_devices.token_hash`.
...
  schemas:
...
    LegacyErrorBody:
      type: object
      description: The ad hoc error shape used by "older" routes (auth, users, badges, points, external-apis). No `status`/`timestamp` field; shape varies slightly per endpoint but always includes `error`.
      required: [error]
      properties:
        error: { type: string }
        message: { type: string }
        required: { type: array, items: { type: string }, description: "Present only on some 'missing required fields' 400s." }
...
    Badge:
      type: object
      description: Raw `badges` table row.
      required: [id, title, category, tier]
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        description: { type: [string, 'null'] }
        icon_emoji: { type: [string, 'null'] }
        category: { type: string }
        tier: { type: string, enum: [bronze, silver, gold, platinum] }
        points_required: { type: [integer, 'null'] }
        created_at: { type: [string, 'null'], format: date-time }
...